move files from admin folder to src folder
