+++ src/components/centerCard/centerCardStyle.js
